simres internal and secondary functions

Listed below are supporting functions for the major methods in simres.

eval.curve evaluates the curve (or expression) expr in the environment specified by enclos.

.cpu.time.ini()

eval.curve(expr, x, enclos = parent.frame())

# S3 method for summaryrhtest
print(x, ...)

Arguments

expr
name of the function (or call or expression written as a function) of x which will evaluate to an object of the same length as x. See for instance curve.

enclos
an enclosure where looks for objects not found in envir.

envir
the environment in which expr is to be evaluated.